requirements.txt
antlr4-python3-runtime==4.11.1
antlr4-tools==0.2
future==0.18.3
install-jdk==0.3.0
numpy==1.24.2
OMPython==3.4.0
psutil==5.9.4
pyparsing==3.0.9
pyzmq==25.0.0
